Update import listings
This updates import listings to make sure that the image is "downloaded" and stored in s3 so that src image displayed on deployed site is our s3 bucket.
Do note to uncomment and comment out the appropriate values when testing locally for s3images. One is using the deployed bucket and another for testing purposes